Add title and chart container components
