+++ src/components/ButtonGroup/ButtonGroup.css
@@ -3,5 +3,4 @@
   height: 30px;
   display: flex;
   justify-content: space-around;
-  background-color: blue;
 }
