+++ src/components/ChartContainer/ChartContainer.css
@@ -1,0 +1,7 @@
+.chart_container {
+  background-color: white;
+  height: 670px;
+  width: 1322px;
+  margin: 0 auto;
+  border-radius: 38px;
+}
