+++ src/components/ChartContainer/ChartContainer.tsx
@@ -1,0 +1,6 @@
+import "./ChartContainer.css";
+import { ReactNode } from "react";
+
+export default function ChartContainer({ children }: { children: ReactNode }) {
+  return <section className="chart_container">{children}</section>;
+}
