+++ src/components/DashBoardTitle/DashboardTitle.css
@@ -1,0 +1,8 @@
+.dash_board_title {
+  font-family: "Roboto";
+  font-weight: bolder;
+  font-size: 25.63px;
+  width: 1322px;
+  height: 30px;
+  margin: 20px auto;
+}
