+++ src/main.tsx
@@ -1,6 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import App from "./App.tsx";
+import "@fontsource/roboto";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
